refactor-#29: adjust application module
