+++ src/customer/application/application.module.ts
@@ -1,4 +1,17 @@
-import { Module } from '@nestjs/common';
+import { Global, Module } from '@nestjs/common';
+import { CqrsModule } from '@nestjs/cqrs';
+import { TypeOrmModule } from '@nestjs/typeorm';
 
-@Module({})
+import { CustomerRepository } from '../repositories';
+import { CommandHandlers } from './commands/handlers';
+import { CustomerController } from './controllers';
+import { QueryHandlers } from './queries/handlers';
+import { CustomerService } from './services';
+
+@Global()
+@Module({
+    imports: [TypeOrmModule.forFeature([CustomerRepository]), CqrsModule],
+    controllers: [CustomerController],
+    providers: [CustomerService, ...CommandHandlers, ...QueryHandlers],
+})
 export class ApplicationModule {}
