+++ src/customer/application/commands/handlers/remove-customer-profile.handler.ts
@@ -3,8 +3,8 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { RpcException } from '@nestjs/microservices';
 
 import { RemoveCustomerProfileCommand } from '../impl';
-import { CustomerRepository } from '../../repositories';
-import { validateDbError } from '../../../database/helpers';
+import { CustomerRepository } from '../../../repositories';
+import { validateDbError } from '../../../../database/helpers';
 
 @CommandHandler(RemoveCustomerProfileCommand)
 export class RemoveCustomerProfileHandler
