+++ src/customer/application/commands/impl/remove-customer-profile.command.ts
