+++ src/customer/application/dto/remove-customer-profile.dto.ts
