+++ src/customer/application/queries/impl/get-customer-profile.query.ts
